Remove Pair
Programmers should never really use classes like "Pair" or "Triplet".
Methods such as #getFirstObject and #getSecondObject convey very little
meaning. It is also unclear which object comes first and why. It is
highly recommended to instead create wrapper classes, when needed, to
bundle two related objects together in your application. The
relationship of the objects in the wrapper class and the method
signatures will be far more readable than using a "Pair" class.